% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doApiQuery.R
\name{convert_list_to_tibble}
\alias{convert_list_to_tibble}
\title{Convert list provided by the APIs into a tibble}
\usage{
convert_list_to_tibble(l)
}
\arguments{
\item{l}{a \link{list} provided by the API (See \link{doApiQuery})}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with one row by list item and one column by list sub-item
}
\description{
Convert list provided by the APIs into a tibble
}
\details{
This function is used internally by all the retrieving data functions for
converting data after the call to \link{doApiQuery}.
}
