% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doQuery.R
\name{doQuery}
\alias{doQuery}
\title{Internal function for doing the queries}
\usage{
doQuery(
  url_path,
  params = list(code_usage = NULL, code_type_eau = NULL, commune_adj = NULL, ecrasant =
    "false"),
  cfg = config::get(file = system.file("config.yml", package = "bnpe")),
  cookie = getCookie(cfg)
)
}
\arguments{
\item{url_path}{the end of the URL path}

\item{params}{A list containing the parameters sent in the query. \link{character} Name of the geographical identifier ("insee_com", "code_dep"...)}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration stored at location \code{system.file("config.yml", package = "bnpe")}}

\item{cookie}{a named \link{character} \link{vector} with the values of the cookies (See \link{getCookie})}
}
\value{
a \link[httr:response]{httr::response} object from a \link[httr:GET]{httr::GET}
}
\description{
Internal function for doing the queries
}
