% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_params.R
\name{get_available_params}
\alias{get_available_params}
\title{List available query parameters of an API endpoint}
\usage{
get_available_params(
  api,
  endpoint,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{api}{a \link{character} name of the API (e.g.: "indicateurs_services", "prelevements"...), see example for available APIs}

\item{endpoint}{a \link{character} name of the endpoint, see example for available endpoints in an API}

\item{cfg}{a \link{config} object describing the configuration of the APIs. Use the internal package configuration by default
configuration}
}
\value{
A \link{character} vector with available query parameters for the requested API/endpoint
}
\description{
List available query parameters of an API endpoint
}
\examples{
# To get the available APIs in the package
cfg <- config::get(file = system.file("config.yml", package = "hubeau"))
names(cfg$apis)

# To get the available endpoints in an API
names(cfg$apis[["prelevements"]]$endpoints)

# To get available parameters in endpoint "chroniques" of the API "prelevements"
get_available_params(api = "prelevements", endpoint = "chroniques")

# To query the endpoint "chroniques" of the API "prelevements"
# on all devices in the commune of Romilly-sur-Seine in 2018
doApiQuery(api = "prelevements",
           endpoint = "chroniques",
           params = list(code_commune_insee = "10323", annee = "2018"))
}
