% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoulement.R
\name{get_ecoulement_stations}
\alias{get_ecoulement_stations}
\alias{get_ecoulement_observations}
\alias{get_ecoulement_campagnes}
\title{Retrieve data from API "Ecoulement des cours d'eau"}
\usage{
get_ecoulement_stations(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_ecoulement_observations(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_ecoulement_campagnes(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{get_available_params} for a list of the available query parameters for a given API endpoint and see the API documentation for their descriptions}

\item{cfg}{a \link{config} object describing the configuration of the APIs. Use the internal package configuration by default
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns.
}
\description{
The data originate from the "ONDE" river low waters monitoring network.
Available endpoints are:
\itemize{
\item \code{get_ecoulement_stations} retrieves site data and locations
\item \code{get_ecoulement_observations} retrieves flow information
\item \code{get_ecoulement_campagnes} retrieves annual surveys
}

See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-ecoulement}
}
\examples{
# Retrieve 2022 observation campaigns in the Jura French department
get_ecoulement_campagnes(
  list(code_departement = "39",
       date_campagne_min = "2022-01-01",
       date_campagne_max = "2022-12-31")
)

# Retrieve river stations
stations_39 <- get_ecoulement_stations(
  list(code_departement = "39",
       fields = "code_station,libelle_cours_eau,libelle_commune")
)
stations_39

# Get the query parameters for the requested API/endpoint
get_available_params(api = "ecoulement",
                     endpoint = "observations")

# Retrieve the river flow data in the Jura departement in 2022 with
# a selection of the fields
onde_39 <- get_ecoulement_observations(
  list(code_departement = "39",
       date_observation_min = "2022-01-01",
       date_observation_max = "2022-12-31",
       fields = "code_station,libelle_station,date_observation,libelle_ecoulement")
)
onde_39

}
\seealso{
\link{doApiQuery} for more details on the API querying process.
}
