% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydrometrie_observations_tr.R
\name{get_hydrometrie_observations_tr}
\alias{get_hydrometrie_observations_tr}
\title{Retrieve hydrometric "real time" observations from API "Hydrométrie"}
\usage{
get_hydrometrie_observations_tr(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{get_available_params} for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns and one row by time step and by station.
}
\description{
See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-hydrometrie}
}
\examples{
# Retrieve the hydrometric last real time observed discharge at station 'H020302002'
get_hydrometrie_observations_tr(list(code_entite = "H0203020", grandeur_hydro = "Q"))

}
