% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydrometrie_stations.R
\name{get_hydrometrie_stations}
\alias{get_hydrometrie_stations}
\title{Retrieve hydrometric stations from API "Hydrométrie"}
\usage{
get_hydrometrie_stations(
  params,
  code_sandre_reseau_station = FALSE,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{get_available_params} for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters}

\item{code_sandre_reseau_station}{optional \link{logical} indicating if \code{code_sandre_reseau_station} field is included in the result; if so, one line is added by item and other fields are repeated}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns and one row by station.
}
\description{
See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-hydrometrie}
}
\examples{
\dontrun{
# This function is temporarily unavailable because of https://github.com/BRGM/hubeau/issues/85

# Retrieve the hydrometric stations in the department of Aube
get_hydrometrie_stations(list(code_departement = "10"))
}
}
