% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicateurs_services.R
\name{get_indicateurs_services_communes}
\alias{get_indicateurs_services_communes}
\alias{get_indicateurs_services_indicateurs}
\alias{get_indicateurs_services_services}
\title{Retrieve performance indicators from drinking water supply and sanitation services at commune level}
\usage{
get_indicateurs_services_communes(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_indicateurs_services_indicateurs(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_indicateurs_services_services(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{get_available_params} for a list of the available query parameters for a given API endpoint and see the API documentation for their description}

\item{cfg}{a \link{config} object describing the configuration of the APIs. Use the internal package configuration by default
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns.
}
\description{
Available endpoints are:
\itemize{
\item 
}

See \link{get_available_params} and the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-indicateurs-services}
}
\details{
\code{get_indicateurs_services_communes} returns a \link[tibble:tibble]{tibble::tibble} with one row by commune, by service and by year and the following columns:
\itemize{
\item "code_commune_insee": \link{character} identifier of the commune
\item "nom_commune": \link{character} name of the commune
\item "codes_service": \link{integer} identifier of the drinking water supply and/or sanitation service
\item "annee": \link{integer} year of the data
\item The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: \url{https://www.services.eaufrance.fr/indicateurs/indicateurs}.
}

a \link[tibble:tibble]{tibble::tibble} with one row by commune, by service and by year and the following columns:
\itemize{
\item "code_commune_insee": \link{character} identifier of the commune
\item "nom_commune": \link{character} name of the commune
\item "codes_service": \link{integer} identifier of the drinking water supply and/or sanitation service
\item "annee": \link{integer} year of the data
\item The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: \url{https://www.services.eaufrance.fr/indicateurs/indicateurs}.
}

\code{get_indicateurs_services_indicateurs} returns a \link[tibble:tibble]{tibble::tibble} with one row by service and by year and the following columns:
\itemize{
\item "code_service": \link{character} identifier of the service
\item "nom_service": \link{character} name of the service
\item "numero_siren ": \link{character} SIREN identifier of the service
\item "type_collectivite": \link{character} kind of collectivity
\item "mode_gestion": \link{character} management mechanism of the service
\item "annee": \link{integer} year of the data
\item "indicateur": value of the indicator
\item "uri_indicateur": the link to the indicator documentation
}

\code{get_indicateurs_services_services} returns a \link[tibble:tibble]{tibble::tibble} with one row by commune, by service and by year and the following columns:
\itemize{
\item "code_service": \link{character} identifier of the service
\item "nom_service": \link{character} name of the service
\item "code_commune_insee": \link{character} identifier of the commune
\item "nom_commune": \link{character} name of the commune
\item "numero_siren ": \link{character} SIREN identifier of the service
\item "type_collectivite": \link{character} kind of collectivity
\item "mode_gestion": \link{character} management mechanism of the service
\item "annee": \link{integer} year of the data
\item The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: \url{https://www.services.eaufrance.fr/indicateurs/indicateurs}.
}
}
\examples{
# Retrieve performance indicator time series in the commune of Romilly-sur-Seine
get_indicateurs_services_communes(list(code_commune = "10323"))

# Retrieve the drinking water withdrawal indicators of the year 2012 for all services
get_indicateurs_services_indicateurs(list(code_indicateur = "D102.0", annee = "2012"))

# Retrieve performance indicator time series of Romilly-sur-Seine with service details
get_indicateurs_services_services(list(code_commune = "10323"))

}
\seealso{
\link{doApiQuery} for more details on the API querying process.
}
