% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicateurs_services_indicateurs.R
\name{get_indicateurs_services_indicateurs}
\alias{get_indicateurs_services_indicateurs}
\title{Retrieve performance indicators from drinking water supply and sanitation services at commune level}
\usage{
get_indicateurs_services_indicateurs(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{\link{list} where the keys are the names of the filtered parameters and the values are the values of the filters. See the API documentation for the complete list of available filter parameters}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with one row by commune, by service and by year and the following columns:
\itemize{
\item "code_commune_insee": \link{character} identifier of the commune
\item "nom_commune": \link{character} name of the commune
\item "codes_service": \link{integer} identifier of the drinking water supply and/or sanitation service
\item "annee": \link{integer} year of the data
\item The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: \url{https://www.services.eaufrance.fr/indicateurs/indicateurs}.
}
}
\description{
See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-indicateurs-services}
}
\examples{
# For retrieving the withdrawal time series of the devices located in Romilly-sur-Seine
get_indicateurs_services_indicateurs(list(code_indicateur = "D102.0", annee = "2012"))

}
