% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_niveaux_nappes_chroniques.R
\name{get_niveaux_nappes_chroniques}
\alias{get_niveaux_nappes_chroniques}
\title{Retrieve piezometric archived time series from API "Piézométrie"}
\usage{
get_niveaux_nappes_chroniques(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{get_available_params} for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns and one row by time step and by station.
}
\description{
See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-piezometrie}
}
\examples{
# Retrieve the archived observed piezometric level at station '07548X0009/F' (old BSS identifier)
# for the year 2020
df <- get_niveaux_nappes_chroniques(list(code_bss = "07548X0009/F",
                                      date_debut_mesure = "2020-01-01",
                                      date_fin_mesure = "2020-12-31"))

# Plot the water elevation (NGF)
plot(as.POSIXct(df$date_mesure), df$niveau_nappe_eau, type = "l")

}
