% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poisson.R
\name{get_poisson_observations}
\alias{get_poisson_observations}
\title{Retrieve data from API "Poisson"}
\usage{
get_poisson_observations(params)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{list_params} for a list of the available filter parameters for a given API endpoint and see the API documentation for their description}
}
\description{
Available endpoint:
\itemize{
\item \code{get_poisson_observations} retrieves data of scientific fishery operations
}

See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-poisson}
}
\examples{
if(interactive()) {
# Get the query parameters for the requested API/endpoint
list_params(api = "poisson",
                     endpoint = "observations")

# Retrieve selected fields on a river fish sampled in Brest
library(dplyr)
fields <- paste("code_operation",
                "date_operation",
                "libelle_point_prelevement_aspe",
                "effectif_lot",
                "code_alternatif_taxon",
                sep = ",")
brest_fishes <- get_poisson_observations(
  list(
    libelle_commune = "Brest",
    fields = fields
    )
 ) \%>\%
 group_by_at(vars(-effectif_lot)) \%>\%
   summarise(nb_individals = sum(effectif_lot))

brest_fishes
}
}
