% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prelevements_chronique.R
\name{get_prelevements_chronique}
\alias{get_prelevements_chronique}
\title{Retrieve time series of withdrawals from Hub'Eau API}
\usage{
get_prelevements_chronique(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{\link{list} where the keys are the names of the filtered parameters and the values are the values of the filters. See the API documentation for the complete list of available filter parameters}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns and one row by device, year and usage.
}
\description{
See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-prelevements-eau#/prelevements/chronique}
}
\examples{
# For retrieving the withdrawal time series of the devices located in Romilly-sur-Seine
get_prelevements_chronique(list(code_commune_insee = "10323"))

}
