% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prelevements_ouvrages.R
\name{get_prelevements_ouvrages}
\alias{get_prelevements_ouvrages}
\title{Retrieve withdrawal devices from API "Prélèvements en eau"}
\usage{
get_prelevements_ouvrages(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{get_available_params} for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns and one row by device.
}
\description{
See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-prelevements-eau}
}
\examples{
# For retrieving the withdrawal devices located in Romilly-sur-Seine
get_prelevements_ouvrages(list(code_commune_insee = "10323"))

}
