% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prelevements_prelevement.R
\name{get_prelevements_prelevement}
\alias{get_prelevements_prelevement}
\title{Retrieve withdrawal points from Hub'Eau API}
\usage{
get_prelevements_prelevement(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{\link{list} where the keys are the names of the filtered parameters and the values are the values of the filters. See the API documentation for the complete list of available filter parameters}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns and one row by device.
}
\description{
See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-prelevements-eau#/prelevements/prelevement}
}
\examples{
# For retrieving the withdrawal points located in Romilly-sur-Seine
get_prelevements_prelevement(list(code_commune_insee = "10323"))

}
