% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualite_nappes.R
\name{get_qualite_nappes_analyses}
\alias{get_qualite_nappes_analyses}
\alias{get_qualite_nappes_stations}
\title{Retrieve data from API "Qualité des nappes d'eau souterraines"}
\usage{
get_qualite_nappes_analyses(params)

get_qualite_nappes_stations(params)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{list_params} for a list of the available filter parameters for a given API endpoint and see the API documentation for their description}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row by record and one column by field.
}
\description{
The data originate from the "ADES" database (\url{https://ades.eaufrance.fr/}).
Available endpoints are:
\itemize{
\item \code{get_qualite_nappes_analyses} retrieves the result analysis of quality measurement
\item \code{get_qualite_eau_potable_resultats_dis} retrieves measuring stations for groundwater quality
}

See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-qualite-nappes}
}
\examples{
\dontrun{
# List of available filter parameters on 'get_qualite_nappe_stations'
list_params("qualite_nappes", "stations")

# List of stations available in 2019 in Hérault department
get_qualite_nappes_stations(list(num_departement = 34,
                                 date_min_maj = "2019-01-01",
                                 date_max_maj = "2019-12-31"))

# List of available filter parameters on 'get_qualite_eau_potable_resultats_dis'
list_params("qualite_nappes", "analyses")

# Get results of analysis realised at Grabels in 2019
get_qualite_nappes_analyses(
  list(bss_id = "BSS002GNSA",
       date_debut_prelevement = "2019-11-12")
)
}
}
