% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComSeries.R
\name{getComSeriesDep}
\alias{getComSeriesDep}
\title{Retrieve data by commune for one departement and one year}
\usage{
getComSeriesDep(
  code_dep,
  year,
  code_usage = NULL,
  code_type_eau = NULL,
  cfg = config::get(file = system.file("config.yml", package = "bnpe")),
  cookie = getCookie(cfg)
)
}
\arguments{
\item{code_dep}{\link{character} of length 2}

\item{year}{\link{numeric} Year of the data to retrieve}

\item{code_usage}{a \code{\link[=character]{character()}} representing the withdrawal destination which can take the following values:
- "ENE" for energy
- "CAN" for artificial canals
- "BAR" for turbined water (dam)
- "AEP" for drinking water
- "IND" for industry
- "IRR" for irrigation}

\item{code_type_eau}{a \code{\link[=character]{character()}} representing the withdrawal source which can take the following values:
- "CONT" for continental surface
- "SOUT" for subsurface
- "LIT" for littoral}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration stored at location \code{system.file("config.yml", package = "bnpe")}}

\item{cookie}{a named \link{character} \link{vector} with the values of the cookies (See \link{getCookie})}
}
\value{
\link{data.frame} with one row by commune and two columns: "insee_com" and "volume"
}
\description{
Retrieve data by commune for one departement and one year
}
\examples{
getComSeriesDep("08", 2016, code_usage = "AEP", code_type_eau = "CONT")
}
