% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCookie.R
\name{getCookie}
\alias{getCookie}
\title{Get cookie from the BNPE website in order to be able to do requests}
\usage{
getCookie(
  cfg = config::get(file = system.file("config.yml", package = "bnpe"))
)
}
\arguments{
\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration stored at location \code{system.file("config.yml", package = "bnpe")}}
}
\value{
a named \link{character} \link{vector} with the values of the cookies
}
\description{
By default it is automatically called at each request but this function
can be used once to store a unique cookie and avoiding asking a cookie for each request (See example).
}
\examples{
cookie <- getCookie()
getTimeSeriesDep("08", cookie = cookie)
getTimeSeriesCom("08105", cookie = cookie)
}
