% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOuvrageSeries.R
\name{getOuvrageSeries}
\alias{getOuvrageSeries}
\title{Retrieve data by device ("ouvrage") (metadata and annual withdrawal volumes)}
\usage{
getOuvrageSeries(
  code_sandre,
  cfg = config::get(file = system.file("config.yml", package = "bnpe")),
  cookie = getCookie(cfg)
)
}
\arguments{
\item{code_sandre}{\link{character} Sandre identifier of the device}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration stored at location \code{system.file("config.yml", package = "bnpe")}}

\item{cookie}{a named \link{character} \link{vector} with the values of the cookies (See \link{getCookie})}
}
\value{
\link{list} compiling data of the device with the following items:
\itemize{
\item \code{id}: BNPE identifier
\item \code{code}: code Sandre
\item \code{codeCommune}: insee code of the commune
\item \code{codePrecision}: ???
\item \code{codeStatut}: ???
\item \code{codeTypeEau}: the withdrawal source which can take the following values:
- "CONT" for continental surface
- "SOUT" for subsurface
- "LIT" for littoral
\item \code{codeUsage}: a numeric code related to water destination: energy, canals, turbined, AEP, industry or irrigation
\item \code{commentaire}: comment
\item \code{exploitationDebut}: start date of the
}
}
\description{
Retrieve data by device ("ouvrage") (metadata and annual withdrawal volumes)
}
\examples{
getOuvrageSeries("OPR0000200109")
}
