% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeries.R
\name{getTimeSeriesCom}
\alias{getTimeSeriesCom}
\alias{getTimeSeriesDep}
\title{Get withdrawal time series for a geographical entity}
\usage{
getTimeSeriesCom(
  insee_com,
  code_usage = NULL,
  code_type_eau = NULL,
  cfg = config::get(file = system.file("config.yml", package = "bnpe")),
  cookie = getCookie(cfg)
)

getTimeSeriesDep(
  insee_dep,
  code_usage = NULL,
  code_type_eau = NULL,
  cfg = config::get(file = system.file("config.yml", package = "bnpe")),
  cookie = getCookie(cfg)
)
}
\arguments{
\item{insee_com}{a \link{character} of length 5 representing the INSEE code of the commune}

\item{code_usage}{a \code{\link[=character]{character()}} representing the withdrawal destination which can take the following values:
- "ENE" for energy
- "CAN" for artificial canals
- "BAR" for turbined water (dam)
- "AEP" for drinking water
- "IND" for industry
- "IRR" for irrigation}

\item{code_type_eau}{a \code{\link[=character]{character()}} representing the withdrawal source which can take the following values:
- "CONT" for continental surface
- "SOUT" for subsurface
- "LIT" for littoral}

\item{cfg}{a \link{config} object Configuration of the communication. Use by default the internal package
configuration stored at location \code{system.file("config.yml", package = "bnpe")}}

\item{cookie}{a named \link{character} \link{vector} with the values of the cookies (See \link{getCookie})}

\item{insee_dep}{a \link{character} of length 2 representing the INSEE code of the commune}
}
\value{
a \link{data.frame} with one year by row and the following columns:
- "volume" the withdrawal annual volume in cubic meters
- "annee" the year
- "peuplement" a unknown code
}
\description{
Get withdrawal time series for a geographical entity
}
\examples{
# Withdrawal time series for Ardennes department
getTimeSeriesDep("08")

# Withdrawal time series for the commune of Charleville-Mézières
getTimeSeriesCom("08105")

}
