% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_obs_elab.R
\name{hm_obs_elab}
\alias{hm_obs_elab}
\title{Get the data corresponding to elaborate observations of a hydrometric station}
\usage{
hm_obs_elab(
  code_entite,
  date_debut,
  date_fin,
  grandeur_hydro_elab = "QmJ",
  variables = c("code_site", "code_station", "date_obs_elab", "resultat_obs_elab",
    "grandeur_hydro_elab")
)
}
\arguments{
\item{code_entite}{the station/entity code}

\item{grandeur_hydro_elab}{the type of hydrometric value ("QmJ" -default- for average daily discharge or "QmM" for average monthly discharge)}

\item{variables}{the list of variables to get as response. Defaults to c("code_site","code_station","date_obs_elab","grandeur_hydro_elab").}
}
\description{
Get the data corresponding to elaborate observations of a hydrometric station
}
\examples{
hm_obs_elab("K340081001",date_debut="2015-12-15",date_fin="2016-01-27")
hm_obs_elab("Y251002001",date_debut="2015-12-15",date_fin="2016-01-27")
}
