% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_obs_elab.R
\name{hm_obs_elab_slice}
\alias{hm_obs_elab_slice}
\title{Get the data corresponding to elaborate observations of a hydrometric station (max 20000 observations)}
\usage{
hm_obs_elab_slice(
  code_entite,
  date_debut,
  date_fin,
  grandeur_hydro_elab = "QmJ"
)
}
\arguments{
\item{code_entite}{the station/entity code}

\item{grandeur_hydro_elab}{the type of hydrometric value ("QmJ" -default- for average daily discharge or "QmM" for average monthly discharge)}

\item{variables}{the list of variables to get as response. Defaults to c("code_site","code_station","date_obs_elab","grandeur_hydro_elab").}
}
\description{
Get the data corresponding to elaborate observations of a hydrometric station (max 20000 observations)
}
