% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_station.R
\name{hm_station}
\alias{hm_station}
\title{Get the information about one hydrometric station}
\usage{
hm_station(code_station)
}
\arguments{
\item{code_station}{the station code}
}
\description{
Get the information about one hydrometric station
}
\examples{
hm_station("K340081001")
}
