% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubeau.R
\docType{package}
\name{hubeau-package}
\alias{hubeau-package}
\title{hubeau: Get Data from the French National Database on Water 'Hub'Eau'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Collection of functions to help retrieving data from 'Hub'Eau' the free and public French National APIs on water \url{https://hubeau.eaufrance.fr/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inrae.github.io/hubeau/}
  \item \url{https://github.com/inrae/hubeau#readme}
  \item Report bugs at \url{https://github.com/inrae/hubeau/issues}
}

}
\author{
\strong{Maintainer}: David Dorchies \email{david.dorchies@inrae.fr} (\href{https://orcid.org/0000-0002-6595-7984}{ORCID})

Other contributors:
\itemize{
  \item Pascal Irz \email{pascal.irz@ofb.gouv.fr} (\href{https://orcid.org/0000-0002-2066-8935}{ORCID}) [contributor]
  \item Sébastien Grall \email{seinormigr.grall@gmail.com} [contributor]
  \item Philippe Amiotte Suchet \email{philippe.amiotte-suchet@u-bourgogne.fr} (\href{https://orcid.org/0000-0003-3514-1447}{ORCID}) [contributor]
}

}
\keyword{internal}
