% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubeau.R
\docType{package}
\name{hubeau}
\alias{hubeau}
\title{hubeau: A package for retrieving data on the French databases on water "Hub'eau"}
\description{
The 'hubeau' package provides functions for APIs and related "operations". These functions are named as follow: \verb{hubeau::get_[API]_[operation]}.
}
\details{
Currently available APIs and related "operations" are listed below.

\strong{API "Hydrométrie"}

API documentation: \url{https://hubeau.eaufrance.fr/page/api-hydrometrie}

Available functions:
\itemize{
\item \link{get_hydrometrie_observations_tr}: hydrometry water level and discharge time series
\item \link{get_hydrometrie_sites}: hydrometry sites (can contain several stations)
\item \link{get_hydrometrie_stations}: hydrometry stations
}

\strong{API "Indicateurs des services"}

API documentation: \url{https://hubeau.eaufrance.fr/page/api-indicateurs-services}

Available functions:
\itemize{
\item \link{get_indicateurs_services_communes}: performance indicators by commune
\item \link{get_indicateurs_services_indicateurs}: performance indicators by indicator
\item \link{get_indicateurs_services_services}: performance indicators by commune for each service
}

\strong{API "Piézométrie"}

API documentation: \url{https://hubeau.eaufrance.fr/page/api-piezometrie}

Available functions:
\itemize{
\item \link{get_niveaux_nappes_chroniques}: archived time series of piezometric stations
\item \link{get_niveaux_nappes_chroniques_tr}: real-time time series of piezometric stations
\item \link{get_niveaux_nappes_stations}: piezometric stations
}

\strong{API "Prélèvements en eau"}

API documentation: \url{https://hubeau.eaufrance.fr/page/api-prelevements-eau}

Available functions:
\itemize{
\item \link{get_prelevements_chroniques}: time series of annual withdrawn volumes by device
\item \link{get_prelevements_ouvrages}: withdrawal devices (can contain several withdrawal points)
\item \link{get_prelevements_points_prelevement}: withdrawal points
}
}
