% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_non_nested}
\alias{select_non_nested}
\title{Selected the non-list elements in a nested list}
\usage{
select_non_nested(my_list)
}
\arguments{
\item{my_list}{A nested list}
}
\value{
The initial nested list omitting its list-class slots
}
\description{
Selected the non-list elements in a nested list
}
\examples{
\dontrun{
l <- doApiQuery(api = "hydrobio",
                endpoint = "stations_hydrobio",
                params = list(code_region = 52))

select_non_nested(l) \%>\%
  select(-code_commune,
         -libelle_departement,
         -(libelle_region:longitude)) \%>\%
  distinct()
}
}
