% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poisson.R
\name{get_poisson_observations}
\alias{get_poisson_observations}
\title{Retrieve data from API "Poisson"}
\usage{
get_poisson_observations(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)
}
\arguments{
\item{params}{a \link{list} the list of parameters of the queries and their values in the format \code{list(ParamName = "Param value", ...)}, use the function \link{get_available_params} for a list of the available query parameters for a given API endpoint and see the API documentation for their descriptions}

\item{cfg}{a \link{config} object describing the configuration of the APIs. Use the internal package configuration by default
configuration}
}
\value{
a \link[tibble:tibble]{tibble::tibble} with all available parameters in columns.
}
\description{
Available endpoint:
\itemize{
\item \code{get_poisson_observations} retrieves data of scientific fishery operations
}

See the API documentation for available filter parameters: \url{https://hubeau.eaufrance.fr/page/api-poisson}
}
\examples{
# Get the query parameters for the requested API/endpoint
get_available_params(api = "poisson",
                     endpoint = "observations")

# Retrieve the river fish sampled in Brest
library(dplyr)
brest_fishes <- get_poisson_observations(
  list(
    libelle_commune = "Brest",
    fields = "code_operation,date_operation,libelle_point_prelevement_aspe,effectif_lot,code_alternatif_taxon"
    )
 ) \%>\%
 group_by_at(vars(-effectif_lot)) \%>\%
   summarise(nb_individals = sum(effectif_lot))

brest_fishes

}
\seealso{
\link{doApiQuery} for more details on the API querying process.
}
