% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubeau.R
\docType{package}
\name{hubeau}
\alias{hubeau}
\title{hubeau: A package for retrieving data on the French databases on water "Hub'eau"}
\description{
The 'hubeau' package provides functions for APIs and related "operations". These functions are named as follow: \verb{hubeau::get_[API]_[operation]}.
}
\details{
Currently available APIs and related "operations" are listed below.

\strong{API "Hydrométrie"}

API documentation: \url{https://hubeau.eaufrance.fr/page/api-hydrometrie}

Available functions:
\itemize{
\item \link{get_hydrometrie_observations_tr}
\item \link{get_hydrometrie_sites}
\item \link{get_hydrometrie_stations}
}

\strong{API "Indicateurs des services"}

API documentation: \url{https://hubeau.eaufrance.fr/page/api-indicateurs-services}

Available functions:
\itemize{
\item \link{get_indicateurs_services_communes}
\item \link{get_indicateurs_services_indicateurs}
}

\strong{API "Prélèvements en eau"}

API documentation: \url{https://hubeau.eaufrance.fr/page/api-prelevements-eau}

Available functions:
\itemize{
\item \link{get_prelevements_chroniques}
\item \link{get_prelevements_ouvrages}
\item \link{get_prelevements_points_prelevement}
}
}
